requirements.txt
setuptools~=65.5.1
mypy==0.990
mypy-extensions==0.4.3
numpy==1.23.4
pandas==1.5.1
gym==0.26.2
docplex==2.23.222
pytest==7.2
cloudpickle==2.2.0
hypothesis==6.56.4